+++ src/backend/resolvers/vega/mutations/index.ts
@@ -3,6 +3,7 @@ import vegaInstitutionResolvers from './institution'
 import vegaPersonResolvers from './person'
 import vegaPersonRoleResolvers from './personRole'
 import vegaPersonRoleTypeResolvers from './personRoleType'
+import vegaProductResolvers from './product'
 import vegaProviderResolvers from './provider'
 
 export default {
@@ -11,5 +12,6 @@ export default {
   ...vegaPersonResolvers,
   ...vegaPersonRoleResolvers,
   ...vegaPersonRoleTypeResolvers,
+  ...vegaProductResolvers,
   ...vegaProviderResolvers,
 }