+++ src/backend/typeDefs/vega/mutations/product/index.ts
@@ -1,0 +1,5 @@
+import updateVegaProductTypeDefs from './update'
+
+export default [
+  updateVegaProductTypeDefs,
+]