+++ src/frontend/Orion/SpecializedData/ValuePerception/ValuePerception.js
@@ -8,9 +8,11 @@ import StakeholderRoles from './StakeholderRoles'
 import Providers from './Providers'
 import Institutions from './Institutions'
 import CommunityPracticeNetworks from './CommunityPracticeNetworks'
+import Products from './Products'
 
 import SANDBOX_MarketBaskets from './SANDBOX_MarketBaskets'
 import SANDBOX_MarketBasketDetail from './SANDBOX_MarketBaskets/MarketBasketDetail'
+import SANDBOX_Products from './SANDBOX_Products'
 import SANDBOX_Stakeholders from './SANDBOX_Stakeholders'
 import SANDBOX_Roles from './SANDBOX_Roles'
 import SANDBOX_RoleSpecialties from './SANDBOX_RoleSpecialties'
@@ -51,6 +53,10 @@ const ValuePerception = () => (
       component={CommunityPracticeNetworks}
     />
     <Route
+      path={'/orion/specialized/value-perception/products'}
+      component={Products}
+    />
+    <Route
       exact={true}
       path={'/orion/specialized/value-perception/sandbox-market-baskets'}
       component={SANDBOX_MarketBaskets}
@@ -60,6 +66,11 @@ const ValuePerception = () => (
         '/orion/specialized/value-perception/sandbox-market-baskets/:marketBasketId'
       }
       component={SANDBOX_MarketBasketDetail}
+    />
+    <Route
+      exact={true}
+      path={'/orion/specialized/value-perception/sandbox-products'}
+      component={SANDBOX_Products}
     />
     <Route
       exact={true}